+++ src/Application/Query/Navigation/SubmenuViewQueryHandler.php
@@ -20,6 +20,7 @@ use Proximum\Vimeet\Application\Query\Navigation\Submenu\ContactsSubmenuViewQuer
 use Proximum\Vimeet\Application\Query\Navigation\Submenu\LeniBadgeLinkSubmenuViewQuery;
 use Proximum\Vimeet\Application\Query\Navigation\Submenu\PackageSubmenuButtonViewQuery;
 use Proximum\Vimeet\Application\Query\Navigation\Submenu\SheetSubmenuViewQuery;
+use Proximum\Vimeet\Application\Query\Navigation\Submenu\UserCtaSubmenuViewQuery;
 use Proximum\Vimeet\Application\Query\Navigation\Submenu\VisioSubmenuViewQuery;
 use Proximum\Vimeet\Application\View\Navigation\SubmenuView;
 
@@ -123,6 +124,19 @@ class SubmenuViewQueryHandler
                 $submenuViewQuery->staticFormulationsIndexByCategories
             )
         );
+
+        $customButtonView = $this->queryBus->handle(
+            new UserCtaSubmenuViewQuery(
+                $submenuViewQuery->user,
+                $submenuViewQuery->event,
+                $submenuViewQuery->locale,
+                $submenuViewQuery->sheet
+            )
+        );
+
+        if (null !== $customButtonView) {
+            $buttonsViews[] = $customButtonView;
+        }
 
         $buttonsViews = array_merge($buttonsViews, $catalogButtonViews);